docs: document server package structure
